requirements.txt
pandas
numpy
scikit-learn
xgboost
joblib
flask
flask-cors
psycopg2-binary
google-generativeai
python-dotenv
